Ignore @base if remote context is not relative
If the remote context is not relative and seems to be a http URI
don't prefix the base IRI to it.

The remote context is not validated nor tested if it's resolvable.
It's just tested if it's not a relative URI - which is totally possible
and would, in conjunction with a base IRI, made into valid remote
context.

See #304.